Mais alguns métodos implementados

+++ src/Structure/CartItem.java
@@ -6,7 +6,7 @@ public class CartItem {
 
 	private Product product;
 	private int reservedQtd;
-	private float curPrice;
+	private double curPrice;
 
 	public CartItem(Product product) {
 		this.product = product;
@@ -16,18 +16,19 @@ public class CartItem {
 	public CartItem(String name, String amount, String price) {
 		product = Products.searchProduct(name);
 		reservedQtd = Integer.parseInt(amount);
-		curPrice = Float.parseFloat(price);
+		curPrice = Double.parseDouble(price);
 	}
 
 	public Product getProduct() { return product; }
 	public String getProductName() { return product.getName(); }
 	public int getReservedQtd() { return reservedQtd; }
 	public String getReservedQtdAsStr() { return Integer.toString(reservedQtd); }
-	public float getPrice() { return curPrice; }
+	public double getPrice() { return curPrice; }
 	public String getPriceAsStr() { return String.format("%.2f", curPrice); }
-	public float getTotalPrice() { return reservedQtd * curPrice; }
+	public double getTotalPrice() { return reservedQtd * curPrice; }
 	public String getTotalPriceAsStr() { return String.format("%.2f", getTotalPrice()); }
 
+	//Adiciona o produto ao carrinho
 	public boolean AddToCart(Cart cart) {
 		if(product.Reserve(1)) {
 			reservedQtd ++;
@@ -37,6 +38,7 @@ public class CartItem {
 		return false;
 	}
 
+	//Atualiza a quantidade solicitada
 	public boolean RefreshQuantity(int qtd) {
 		if (product.Reserve(qtd)) {
 			reservedQtd += qtd;
@@ -45,6 +47,7 @@ public class CartItem {
 		return false;
 	}
 
+	//Remove do carrinho
 	public void RemoveFromCart(int qtd) {
 		product.CancelReservation(qtd);
 		reservedQtd -= qtd;
